+++ setup.cfg
@@ -4,6 +4,5 @@ exclude =
     .git,
     __pycache__,
     baselines/her,
-    baselines/ddpg,
     baselines/ppo1,
     baselines/bench,
